% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSigma.R
\name{FitSigma}
\alias{FitSigma}
\title{Seed viability curve fitting to estimate
\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}}}
\usage{
FitSigma(data, viability.percent, samp.size, storage.period,
  probit.method = c("glm", "tflm"))
}
\arguments{
\item{data}{A data frame with the seed viability data recorded periodically.
It should possess columns with data on \itemize{ \item Viability percentage
(to be indicated by the argument \code{viability.percent}), \item Sample
size (to be indicated by the argument \code{samp.size}) and \item Storage
period (to be indicated by the argument \code{storage.period}). }}

\item{viability.percent}{The name of the column in \code{data} with the
viability percentages as a character string.}

\item{samp.size}{The name of the column in \code{data} with the sample size
used for calculating viability percentages as a character string.}

\item{storage.period}{The name of the column in \code{data} with the time
periods at which the viabilty percentages was recorded as a character
string.}

\item{probit.method}{The method to be used for fitting seed viability curve.
Either as a generalised linear model with a probit link function
(\code{"glm"}, recommended) or as a linear model with probit transformed
viability percentages (\code{"tflm"}).}
}
\value{
A list with the following components:  \item{model}{The
  fitted model as an object of class \code{glm} (if \code{probit.method =
  "glm"}) or \code{lm} (if \code{probit.method = "lm"}).} \item{parameters}{A
  data.frame of parameter estimates, standard errors and p value.}
  \item{fit}{A one-row data frame with estimates of model fitness such as log
  likelyhoods, Akaike Information Criterion, Bayesian Information Criterion,
  deviance and residual degrees of freedom.} \item{Ki}{The estimated seed lot
  constant from the model.} \item{sigma}{The estimated period of time to lose
  unit probit viability from the model.} \item{message}{Warning or error
  messages generated during fitting of model, if any.}
}
\description{
Fit seed viability/survival curve to estimate the seed lot constant
(\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}}) and the period to
lose unit probit viability
(\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}}).
}
\details{
This function fits seed survival data to the following seed viability
equation \insertCite{ellis_improved_1980}{viabilitymetrics} which models the
relationship between probit percentage viability and time period of storage.

\ifelse{html}{\out{<p style="text-align: center;"><em>v = K<sub>i</sub>
&minus; [ <sup>p</sup> &frasl; <sub>&sigma;</sub> ]
</em></p>}}{\deqn{v=K_{i}-\frac{p}{\sigma}}}

or

\ifelse{html}{\out{<p style="text-align: center;"><em>v = K<sub>i</sub>
&minus; <big>(</big><sup>1</sup> &frasl;
<sub>&sigma;</sub><big>)</big>&sdot;p</em></p>}}{\deqn{v=K_{i}-\left (
\frac{1}{\sigma} \right )\cdot p}}

Where, \ifelse{html}{\out{<i>v</i>}}{\eqn{v}} is the probit percentage
viability at storage time \ifelse{html}{\out{<i>p</i>}}{\eqn{p}} (final
viability),  \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} is the
probit percentage viability of the seedlot at the beginning of storage
(seedlot constant) and
\ifelse{html}{\out{<em><sup>1</sup>&frasl;<sub>&sigma;</sub></em>}}{\eqn{\frac{1}{\sigma}}}
is the slope.

The above equation may be expressed as a generalized linear model (GLM) with
a probit (cumulative normal distribution) link function as follows
\insertCite{hay_modelling_2014}{viabilitymetrics}.

\ifelse{html}{\out{<p style="text-align: center;"><em>y = &phi;(v) =
&phi;<big>(</big>K<sub>i</sub> &minus; (<sup>1</sup> &frasl;
<sub>&sigma;</sub>) <big>)</big></em></p>}}{\deqn{y = \phi(v) = \phi\left (
K_{i}-\left ( \frac{1}{\sigma} \right )p \right )}}

Where, \ifelse{html}{\out{<i>y</i>}}{\eqn{y}} is the proportion of seeds
viabile after time period \ifelse{html}{\out{<i>p</i>}}{\eqn{p}} and the link
function is
\ifelse{html}{\out{<em>&phi;<sup>-1</sup></em>}}{\eqn{\phi^{-1}}}, the
inverse of the cumulative normal distribution function.

The parameters estimated are the intercept
\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K<i>}}, theoretical viability
of the seeds at the start of storage or the seed lot constant, and the slope
\ifelse{html}{\out{<i>&minus;&sigma;<sup>-1</sup></i>}}{\eqn{-\sigma^{-1}}},
where \ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} is the standard
deviation of the normal distribution of seed deaths in time or the period of
time to lose unit probit viability.
}
\examples{
df <- data.frame(Rep = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
                         2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2),
                 period = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
                            11, 12, 13, 14, 15, 1, 2, 3, 4, 5,
                            6, 7, 8, 9, 10, 11, 12, 13, 14, 15),
                 sampsize = c(97, 100, 100, 100, 98, 100, 100, 100,
                              99, 98, 101, 100, 100, 101, 100, 100,
                              99, 97, 100, 97, 99, 98, 100, 98, 99,
                              99, 101, 98, 97, 97),
                 viabilitypercent = c(98, 96, 91, 83, 70, 55, 39, 24,
                                      13, 7, 3, 1, 0, 0, 0, 98, 94, 86,
                                      73, 55, 35, 19, 9, 3, 1,
                                      0, 0, 0, 0, 0))

plot(df$period, df$viabilitypercent)

#----------------------------------------------------------------------------
# With generalised linear model with probit link function
#----------------------------------------------------------------------------
model1 <- FitSigma(data = df, viability.percent = "viabilitypercent",
                   samp.size = "sampsize", storage.period = "period",
                   probit.method = "glm")
model1
# Raw model
model1$model

# Model parameters
model1$parameters

# Model fit
model1$fit

#----------------------------------------------------------------------------
# With linear model after probit transformation
#----------------------------------------------------------------------------
model2 <- FitSigma(data = df, viability.percent = "viabilitypercent",
                   samp.size = "sampsize", storage.period = "period",
                   probit.method = "tflm")
model2
# Raw model
model2$model

# Model parameters
model2$parameters

# Model fit
model2$fit

}
\references{
\insertAllCited{}
}
