% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ke.R
\encoding{UTF-8}
\name{Ke}
\alias{Ke}
\title{Viability constant \mjseqn{K_{E}} based on the species-specific temperature
coefficients}
\usage{
Ke(K, temp, temp.coeff = c(0.0329, 0.000478))
}
\arguments{
\item{K}{The constant \mjseqn{K} associated with the relationship of
temperature with seed longevity (see \strong{Details}).}

\item{temp}{Temperature in °C.}

\item{temp.coeff}{The species-specific temperature coefficients
(\mjseqn{C_{H}} and \mjseqn{C_{Q}}) as a numeric vector of length 2.}
}
\value{
The value of species-specific seed viability constant \mjseqn{K_{E}}.
}
\description{
\code{Ke} computes the viability constant \mjseqn{K_{E}} from the
species-specific temperature coefficients in case of storage experiments
conducted at constant temperature and varying moisture contents. \loadmathjax
}
\details{
From seed storage experiments involving storage of seeds at a constant
temperature in a range of moisture contents, the effect of moisture content
on seed longevity (\mjseqn{\sigma}) can be estimated from the following
linear relationship:

\mjsdeqn{\log{\sigma} = K - C_{w}\log{m}}

Where, \mjseqn{K} is the intercept, \mjseqn{C_{W}} is the slope and
\mjseqn{m} is the moisture content.

The constant \mjseqn{K} associated with the relationship of temperature with
seed longevity as follows.

\mjsdeqn{K = K_{E} - C_{H}t - C_{Q}t^{2}}

Where, \mjseqn{K_{E}}, \mjseqn{C_{H}} and \mjseqn{C_{Q}} are the
species-specific seed viability constants.

The constant \mjseqn{K_{E}} can be estimated from the universal temperature
constants (\mjseqn{C_{H}} = 0.0329 and \mjseqn{C_{Q}} = 0.000478) in case of
seed storage experiments at constant temperature and varying moisture content
as follows.

\mjsdeqn{K_{E} = K + C_{H}t + C_{Q}t^{2}}
}
\examples{
Ke(36, 10)

}
\references{
\insertRef{ellis_improved_1980}{viabilitymetrics}

\insertRef{pritchard_predicting_2003}{viabilitymetrics}
}
\seealso{
\code{\link[viabilitymetrics]{Sigma}}
}
