% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p50.R
\encoding{UTF-8}
\name{P50}
\alias{P50}
\title{Half-viability period}
\usage{
P50(initial, vcindex, vcdirect, mc, temp, years = FALSE)
}
\arguments{
\item{initial}{The initial viability (\%).}

\item{vcindex}{An interger value indicating the index of seed viability.
constants to be used from the \code{viabilityconstants} dataset in the
package.}

\item{vcdirect}{A numeric vector of length 4 with the four viability
constants \emph{viz.}:
\ifelse{html}{\out{<i>K<sub>E</sub></i>}}{\eqn{K_{E}}},
\ifelse{html}{\out{<i>C<sub>W</sub></i>}}{\eqn{C_{W}}},
\ifelse{html}{\out{<i>C<sub>H</sub></i>}}{\eqn{C_{H}}} and
\ifelse{html}{\out{<i>C<sub>Q</sub></i>}}{\eqn{C_{Q}}}.}

\item{mc}{Moisture content.}

\item{temp}{Temperature in °C.}

\item{years}{If \code{TRUE}, returns the output period in years instead of
days.}
}
\value{
The half-viability period in days or years (according to argument
  \code{years}).
}
\description{
\code{P50} computes the half-viability period, which is the time taken for
50\% of the seeds to lose viability.
}
\details{
The period to lose 50\% viability
(\ifelse{html}{\out{<em>P<sub>50</sub></em>}}{\eqn{P_{50}}}) is computed
according to the relationship between probit percentage viabilities and time
of storage described by Ellis and Roberts (1980) as follows:

\ifelse{html}{\out{<p style="text-align: center;"><em>v = K<sub>i</sub>
&minus; [ <sup>p</sup> &frasl; <sub>&sigma;</sub> ]
</em></p>}}{\deqn{v=K_{i}-\frac{p}{\sigma}}}

or

\ifelse{html}{\out{<p style="text-align: center;"><em>v = K<sub>i</sub>
&minus; <big>(</big><sup>1</sup> &frasl;
<sub>&sigma;</sub><big>)</big>&sdot;p</em></p>}}{\deqn{v=K_{i}-\left (
\frac{1}{\sigma} \right )\cdot p}}

Where, \ifelse{html}{\out{<i>v</i>}}{\eqn{v}} is the probit percentage
viability at storage time \ifelse{html}{\out{<i>p</i>}}{\eqn{p}} (final
viability),  \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} is the
probit percentage viability of the seedlot at the beginning of storage
(seedlot constant) and
\ifelse{html}{\out{<em><sup>1</sup>&frasl;<sub>&sigma;</sub></em>}}{\eqn{\frac{1}{\sigma}}}
is the slope.

Germination percentages plotted against storage times yield a sigmoid seed
survival curve which is converted to a linear relationship by the probit
transformation with slope \ifelse{html}{\out{<i><sup>1</sup> &frasl;
<sub>&sigma;</sub></i>}}{\eqn{\frac{1}{\sigma}}}.

When \ifelse{html}{\out{<i>v</i>}}{\eqn{v}} = 0 (equivalent to 50\%
viability), \ifelse{html}{\out{<em>P<sub>50</sub></em>}}{\eqn{P_{50}}} can be
computed as follows:

\ifelse{html}{\out{<p style="text-align: center;"><em>P<sub>50</sub> =
K<sub>i</sub> &times; &sigma;</em></p>}}{\deqn{P_{50} = K_{i} \times \sigma}}

If the initial viablity (\code{initial}) is beyond limits (0-100 \%, an error
is issued.

The value of this computation is appropriate for temperature between -20 to
90 °C and seed moisture content between 5 to 25\%. For values beyond this
range, a warning will be displayed.
}
\note{
For initial viability percentage values of 100\%, adjust it according
  to sample size using the
  \code{\link[viabilitymetrics:Percent2Probit]{PercentAdjust}} function to
  avoid infinity values in output.
}
\examples{
P50(initial = 98, vcindex = 24, mc = 5, temp = -20)
P50(initial = 98, vcindex = 24, mc = 5, temp = -20, years = TRUE)

# With intial viability of 100\%
P50(initial = 100, vcindex = 24, mc = 5, temp = -20)
P50(initial = 100, vcindex = 24, mc = 5, temp = -20, years = TRUE)

# With intial viability of 100\%, use of PercentAdjust() to avoid Inf
P50(initial = PercentAdjust(100, n = 50), vcindex = 24, mc = 5, temp = -20)
P50(initial = PercentAdjust(100, n = 50), vcindex = 24, mc = 5, temp = -20,
    years = TRUE)

\dontrun{
# Error if initial viability is beyond limits (0-100 \%)
P50(initial = 110, vcindex = 24, mc = 5, temp = -20)
}

}
\references{
\insertRef{ellis_improved_1980}{viabilitymetrics}
}
\seealso{
\code{\link[viabilitymetrics]{Sigma}},
  \code{\link[viabilitymetrics:Percent2Probit]{PercentAdjust}}
}
