% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent2probit.R
\name{Percent2Probit}
\alias{Percent2Probit}
\alias{PercentAdjust}
\alias{Percent2NED}
\alias{Probit2NED}
\alias{NED2Probit}
\alias{NED2Percent}
\alias{Probit2Percent}
\title{Probit transformation}
\usage{
PercentAdjust(percentage, n)

Percent2NED(percentage)

Percent2Probit(percentage)

Probit2NED(probit)

NED2Probit(NED)

NED2Percent(NED)

Probit2Percent(probit)
}
\arguments{
\item{percentage}{The percentage value.}

\item{n}{Sample size for estimation of percentage.}

\item{probit}{The probit value}

\item{NED}{The NED value.}
}
\value{
The transformed value.
}
\description{
These functions transform data between percentage, probit and Normal
Equivalent Deviate (NED)
\insertCite{bliss_method_1934,finney_probit_1952}{viabilitymetrics}.
}
\details{
Probit transformation can be used to transform a sigmoid curve of percentage
data to a linear one. The probit transformation is defined as
\ifelse{html}{\out{NED + 5}}{\eqn{NED + 5}}. However the two terms probit and
NED are used interchangeably in literature.

NED function
(\ifelse{html}{\out{<i>&Phi;<sup>-1</sup></i>}}{\eqn{\Phi^{-1}}}) is is the
inverse of the cumulative distribution function
(\ifelse{html}{\out{<i>&Phi;</i>}}{\eqn{\Phi}}) of the standard normal
distribution (\ifelse{html}{\out{<em>z &sim; N(0,1)</em>}}{\eqn{z\sim
N(0,1)}}) or the quantile function associated with the standard normal
distribution.

For percentage \ifelse{html}{\out{<i>p</i>}}{\eqn{p}},

\ifelse{html}{\out{<p style="text-align: center;">NED<em>(p)=
&Phi;<sup>-1</sup>(p) = &radic;<span style="text-decoration:
overline;">2</span></em> erf<sup>-1</sup><em>
(2p-1)</em></p>}}{\deqn{\textrm{NED}(p)=\Phi^{-1}(p)=
\sqrt{2}~\textrm{erf}^{-1}(2p-1)}}

and

\ifelse{html}{\out{<p style="text-align: center;">probit<em>(p) =</em>
NED<em>(p) + 5</em></p>}}{\deqn{\textrm{probit}(p)=\textrm{NED}(p)+5}}

The \code{PercentAdjust} function adjusts the percentage values of 0 and 100
to \ifelse{html}{\out{100 &times; [ <sup>0.25</sup> &frasl;
<sub><i>n</i></sub> ]}}{\eqn{100\times \frac{0.25}{n}}} and
\ifelse{html}{\out{100 &times; [ <sup>(<i>n</i> &minus; 0.25)</sup> &frasl;
<sub><i>n</i></sub> ]}}{\eqn{100\times \frac{n-0.25}{n}}} respectively,
according to the sample size \ifelse{html}{\out{<i>n</i>}}{\eqn{n}} to avoid
infinity values during probit transformation
\insertCite{miller_estimation_1944}{viabilitymetrics}.
}
\examples{
Percent2NED(0:100)
Percent2Probit(0:100)

Percent2NED(25)
Percent2Probit(25)
Percent2NED(25) +5
NED2Probit(-0.6744898)

# Percentage adjustment for 0 and 100
Percent2Probit(100)
Percent2Probit(0)
n = 50
Percent2Probit(PercentAdjust(100, n))
Percent2Probit(PercentAdjust(0, n))

}
\references{
\insertRef{bliss_method_1934}{viabilitymetrics}

  \insertRef{miller_estimation_1944}{viabilitymetrics}

  \insertRef{finney_probit_1952}{viabilitymetrics}
}
