% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FitSigma.batch.R
\name{plot.FitSigma.batch}
\alias{plot.FitSigma.batch}
\title{Plot the fitted seed viability curves from a \code{FitSigma.batch} object}
\usage{
\method{plot}{FitSigma.batch}(x, limits = TRUE, grid = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{FitSigma.batch} obtained as output from the
\code{\link[viabilitymetrics]{FitSigma.batch}} function.}

\item{limits}{logical. If \code{TRUE}, set the limits of y axis (viability
percentage) between 0 and 100 in the viability curve plot. If
\code{FALSE}, limits are set according to the data. Default is \code{TRUE}.}

\item{grid}{logical. If \code{TRUE}, a symmetric matrix grid of plots is
produced instead of a single plot with multiple curves. Default is
\code{FALSE}.}

\item{\dots}{Default plot arguments.}
}
\value{
The plot of the seed viability curves as an object of class
  \code{ggplot}.
}
\description{
\code{plot.FitSigma.batch} plots the group-wise fitted seed
viability/survival curves from a \code{FitSigma.batch} object as an object of
class \code{ggplot}.
}
\examples{

data(seedsurvival)
df <- seedsurvival[seedsurvival$moistruecontent == 7 &
                     seedsurvival$temperature == 25,
                   c("crop", "storageperiod", "rep",
                     "viabilitypercent", "sampsize")]

#----------------------------------------------------------------------------
# Generalised linear model with probit link function (without cv)
#----------------------------------------------------------------------------
model1a <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize",
                          storage.period = "storageperiod",
                          probit.method = "glm")
plot(model1a)
plot(model1a, grid = TRUE)

#----------------------------------------------------------------------------
# Generalised linear model with probit link function (with cv)
#----------------------------------------------------------------------------
model1b <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize",
                          storage.period = "storageperiod",
                          probit.method = "glm",
                          use.cv = TRUE, control.viability = 98)
plot(model1b)
plot(model1b, grid = TRUE)

#----------------------------------------------------------------------------
# Linear model after probit transformation (without cv)
#----------------------------------------------------------------------------
model2a <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize",
                          storage.period = "storageperiod",
                          probit.method = "tflm")
plot(model2a)
plot(model2a, grid = TRUE)

#----------------------------------------------------------------------------
# Linear model after probit transformation (with cv)
#----------------------------------------------------------------------------
model2b <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize",
                          storage.period = "storageperiod",
                          probit.method = "tflm",
                          use.cv = TRUE, control.viability = 98)
plot(model2b)
plot(model2b, grid = TRUE)

}
\seealso{
\code{\link[viabilitymetrics]{FitSigma.batch}}
}
