% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FitSigma.R
\name{plot.FitSigma}
\alias{plot.FitSigma}
\title{Plot the fitted seed viability curve from a \code{FitSigma} object}
\usage{
\method{plot}{FitSigma}(x, limits = TRUE, annotate = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{FitSigma} obtained as output from the
\code{\link[viabilitymetrics]{FitSigma}} function.}

\item{limits}{logical. If \code{TRUE}, set the limits of y axis (viability
percentage) between 0 and 100 in the viability curve plot. If
\code{FALSE}, limits are set according to the data. Default is \code{TRUE}.}

\item{annotate}{logical. If \code{TRUE},
\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} values are annotated on
the plot. Default is \code{TRUE}.}

\item{\dots}{Default plot arguments.}
}
\value{
The plot of the seed viability curve as an object of class
  \code{ggplot}.
}
\description{
\code{plot.FitSigma} plots the fitted seed viability/survival curve from a
\code{FitSigma} object as an object of class \code{ggplot}.
}
\examples{

data(seedsurvival)
df <- seedsurvival[seedsurvival$crop == "Soybean" &
                     seedsurvival$moistruecontent == 7 &
                     seedsurvival$temperature == 25,
                   c("storageperiod", "rep",
                     "viabilitypercent", "sampsize")]

#----------------------------------------------------------------------------
# Generalised linear model with probit link function (without cv)
#----------------------------------------------------------------------------
model1a <- FitSigma(data = df, viability.percent = "viabilitypercent",
                   samp.size = "sampsize", storage.period = "storageperiod",
                   generalised = TRUE)
plot(model1a)

#----------------------------------------------------------------------------
# Generalised linear model with probit link function (with cv)
#----------------------------------------------------------------------------
model1b <- FitSigma(data = df, viability.percent = "viabilitypercent",
                   samp.size = "sampsize", storage.period = "storageperiod",
                   generalised = TRUE,
                   use.cv = TRUE, control.viability = 98)
plot(model1b)

#----------------------------------------------------------------------------
# Linear model after probit transformation (without cv)
#----------------------------------------------------------------------------
model2a <- FitSigma(data = df, viability.percent = "viabilitypercent",
                   samp.size = "sampsize", storage.period = "storageperiod",
                   generalised = FALSE)
plot(model2a)

#----------------------------------------------------------------------------
# Linear model after probit transformation (with cv)
#----------------------------------------------------------------------------
model2b <- FitSigma(data = df, viability.percent = "viabilitypercent",
                   samp.size = "sampsize", storage.period = "storageperiod",
                   generalised = FALSE,
                   use.cv = TRUE, control.viability = 98)
plot(model2b)

}
\seealso{
\code{\link[viabilitymetrics]{FitSigma}}
}
