% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FitSigma.batch.R
\name{print.FitSigma.batch}
\alias{print.FitSigma.batch}
\title{Prints \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} from a
\code{FitSigma.batch} object}
\usage{
\method{print}{FitSigma.batch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{print.FitSigma.batch}.}

\item{...}{Unused}
}
\value{
The \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
  \ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} values (degree Celsius)
  in the console.
}
\description{
\code{print.FitSigma.batch} prints to console the seed lot constant
(\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}}) and the period to
lose unit probit viability
(\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}}).
}
\seealso{
\code{\link[viabilitymetrics]{FitSigma.batch}}
}
