% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FitSigma.R
\name{print.FitSigma}
\alias{print.FitSigma}
\title{Prints \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} from a \code{FitSigma}
object}
\usage{
\method{print}{FitSigma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{print.FitSigma}.}

\item{...}{Unused}
}
\value{
The \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
  \ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} values (degree Celsius) in
  the console.
}
\description{
\code{print.FitSigma} prints to console the seed lot constant
(\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}}) and the period to
lose unit probit viability
(\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}}).
}
\seealso{
\code{\link[viabilitymetrics]{print.FitSigma}}
}
