% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedEqMC.R
\encoding{UTF-8}
\name{SeedEqMC}
\alias{SeedEqMC}
\alias{EqRH}
\title{Seed equilibrium moisture content and equilibrium relative humidity of the
seed storage environment}
\usage{
SeedEqMC(oilcontent, rh, temp, basis = c("wet", "dry"))

EqRH(oilcontent, mc, temp, basis = c("wet", "dry"))
}
\arguments{
\item{oilcontent}{The percentage oil content of seed (dry basis).}

\item{rh}{Relative humidity expressed in percentage.}

\item{temp}{Temperature in 째C.}

\item{basis}{The type of estimation of moisture content specified in the
argument \code{mc}. Either \code{"wet"} or \code{"dry"}.}

\item{mc}{The seed equilibrium moisture content on wet or dry basis
(according to argument \code{basis}).}
}
\value{
For \code{SeedEqMC}, the seed equilibrium moisture content on wet or
  dry basis (according to argument \code{basis}) expressed in percentage.

  For \code{EqRH}, the equilibrium relative humidity expressed in percentage.
}
\description{
Compute the following metrics:\describe{\item{\code{SeedEqMC}}{The seed
equilibrium moisture content from known environmental conditions and oil
content.} \item{\code{EqRH}}{The equilibrium relative humidity of the seed
storage environment from seed equilibrium moisture content, oil content and
temperature.} } \loadmathjax
}
\details{
This relationship between seed equilibrium moisture content, seed oil
content, the  equilibrium relative humidity and temperature of the storage
environment was described by
\insertCite{cromarty_design_1982;textual}{viabilitymetrics} as follows.

\mjsdeqn{(1-R) = \textrm{e}^{-\left( {\bigg[ \frac{M_{e} \times \big(1.1 +
\frac{T}{90} \big)}{1 - D_{O}} \bigg]^{2}}\bigg/{440} \right)}}

Where, \mjseqn{R} is the relative humidity expressed as decimal,
\mjseqn{M_{e}} is the equilibrium percentage moisture content (dry basis),
\mjseqn{T} is the temperature in 째C of air or the seed equilibrium,
\mjseqn{D_{O}} is the oil content of seed (dry basis) expressed as decimal
and \mjseqn{e} is the mathematical constant 2.718282.

For values of oil content (\code{oilcontent}), relative humidity (\code{rh})
and seed equilibrium moisture content (\code{mc}) beyond the limits of 0-100
\%, a warning is issued.
}
\note{
The above expression by
  \insertCite{cromarty_design_1982;textual}{viabilitymetrics} is recommended
  for temperature and humidity ranges of 0-40 째C and 10-70\% RH for starchy
  seeds (eg. cereals); and 15-25 째C and 10-70\% RH for oilseeds.
}
\examples{
SeedEqMC(oilcontent = 29, rh = 13, temp = 25, basis = "wet")
SeedEqMC(oilcontent = 29, rh = 13, temp = 25, basis = "dry")

EqRH(oilcontent = 29, mc = 5, temp = 25, basis = "wet")
EqRH(oilcontent = 29, mc = 5, temp = 25, basis = "dry")

# Warning if oilcontent is beyond limits (0-100 \%)
SeedEqMC(oilcontent = 125, rh = 13, temp = 25, basis = "wet")
EqRH(oilcontent = 125, mc = 5, temp = 25, basis = "wet")

# Warning if relative humidity is beyond limits (0-100 \%)
SeedEqMC(oilcontent = 29, rh = 115, temp = 25, basis = "wet")

# Warning if moisture content is beyond limits (0-100 \%)
EqRH(oilcontent = 29, mc = 115, temp = 25, basis = "wet")

}
\references{
\insertAllCited
}
\seealso{
\code{\link[viabilitymetrics]{wet2dry}}
}
