% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma.R
\encoding{UTF-8}
\name{Sigma}
\alias{Sigma}
\title{Period to lose unit probit viability}
\usage{
Sigma(vcindex, vcdirect, mc, temp, years = FALSE)
}
\arguments{
\item{vcindex}{An integer value indicating the index of seed viability.
constants to be used from the \code{viabilityconstants} dataset in the
package.}

\item{vcdirect}{A numeric vector of length 4 with the four viability
constants \emph{viz.}: \mjseqn{K_{E}}, \mjseqn{C_{W}}, \mjseqn{C_{H}} and
\mjseqn{C_{Q}}.}

\item{mc}{Moisture content.}

\item{temp}{Temperature in °C.}

\item{years}{If \code{TRUE}, returns the output period in years instead of
days.}
}
\value{
The period to lose one probit in days or years (according to argument
  \code{years}).
}
\description{
\code{Sigma} calculates the period to lose one probit viability
(\mjseqn{\sigma}) under storage at a given moisture content and temperature.
\loadmathjax
}
\details{
This function computes the period to lose one probit viability
(\mjseqn{\sigma}) according to the improved seed viability equation of
\insertCite{ellis_improved_1980;textual}{viabilitymetrics} as follows.

\mjsdeqn{v=K_{i}-\frac{p}{\sigma}}

or

\mjsdeqn{v=K_{i}-\left ( \frac{1}{\sigma} \right )\cdot p}

Where, \mjseqn{v} is the probit percentage viability at storage time
\mjseqn{p} (final viability),  \mjseqn{K_{i}} is the probit percentage
viability of the seedlot at the beginning of storage (seedlot constant) and
\mjseqn{\frac{1}{\sigma}} is the slope.

Germination percentages plotted against storage times yield a sigmoid seed
survival curve which is converted to a linear relationship by the probit
transformation with slope \mjseqn{\frac{1}{\sigma}}.

The slope is determined as follows.

\mjsdeqn{\sigma = 10^{K_{E}-C_{W}\log{m}-C_{H}t-C_{Q}t^2}}

Where, \mjseqn{v} is the probit percentage viability at storage time
\mjseqn{p} (final viability), \mjseqn{K_{i}} is the probit percentage
viability of the seedlot at the beginning of storage (seedlot constant),
\mjseqn{m} is the moisture content (fresh weight basis), \mjseqn{t} is the
temperature and \mjseqn{K_{E}}, \mjseqn{C_{W}}, \mjseqn{C_{H}} and
\mjseqn{C_{Q}} are the species-specific seed viability constants.

The value of the species-specific seed viability constants can be specified
either directly in the argument \code{vcdirect} or as the index value of the
required seed viability constants from the \code{\link{viabilityconstants}}
dataset through the argument \code{vcindex}.

The value of this prediction is appropriate for temperature between -20 to 90
°C and seed moisture content between 5 to 25\%. For values beyond this range,
a warning will be displayed.
}
\examples{
#----------------------------------------------------------------------------
# Days/Years to lose unit probit viability for rice seeds stored at
# 5 degree celsius and 10\% moisture content.
#----------------------------------------------------------------------------
# Fetch the index from viabilityconstants dataset
viabilityconstants[grepl("oryza", x = viabilityconstants$Species,
                         ignore.case = TRUE),]
# Use index 87
Sigma(vcindex = 87, mc = 10, temp = 5)
Sigma(vcindex = 87, mc = 10, temp = 5, years = TRUE)

# Input the viability constants directly
Sigma(vcdirect = c(8.242, 4.345, 0.0307, 0.000501), mc = 10, temp = 5)
Sigma(vcdirect = c(8.242, 4.345, 0.0307, 0.000501), mc = 10, temp = 5,
      years = TRUE)

# Warning if moisture content is beyond limits (0-100 \%)
Sigma(vcindex = 87, mc = 110, temp = 5)

# Warning if temperature is beyond limits (-20 to 90 degree C)
Sigma(vcindex = 87, mc = 10, temp = 95)

#----------------------------------------------------------------------------
# Days/Years to lose unit probit viability for soybean seeds stored at
# -18 degree celsius and 8\% moisture content.
#----------------------------------------------------------------------------
# Fetch the index from viabilityconstants dataset
viabilityconstants[grepl("glycine", x = viabilityconstants$Species,
                         ignore.case = TRUE),]
# Use index  59
Sigma(vcindex = 59, mc = 8, temp = -18)
Sigma(vcindex = 59, mc = 8, temp = -18, years = TRUE)

# Input the viability constants directly
Sigma(vcdirect = c(7.292, 3.996, 0.0295, 0.000491), mc = 8, temp = -18)
Sigma(vcdirect = c(7.292, 3.996, 0.0295, 0.000491), mc = 8, temp = -18,
      years = TRUE)

# Warning if moisture content is beyond limits (0-100 \%)
Sigma(vcindex = 59, mc = 110, temp = 5)

# Warning if temperature is beyond limits (-20 to 90 degree C)
Sigma(vcindex = 59, mc = 10, temp = 95)

}
\references{
\insertAllCited
}
