% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaTransformed.R
\encoding{UTF-8}
\name{SigmaTransformed}
\alias{SigmaTransformed}
\title{SigmaTransformed}
\usage{
SigmaTransformed(sigma, temp1, temp2, temp.coeff = c(0.0329, 0.000478))
}
\arguments{
\item{sigma}{The inverse of slope from the seed viability equation
(\mjseqn{\sigma}), estimated at temperature \code{temp1}.}

\item{temp1}{The temperature at which \code{sigma} is estimated in °C.}

\item{temp2}{The temperature at which the transformed \mjseqn{\sigma} is to
be estimated in °C.}

\item{temp.coeff}{The species-specific temperature coefficients
(\mjseqn{C_{H}} and \mjseqn{C_{Q}}.) as a numeric vector of length 2.}
}
\value{
The transformed value of \mjseqn{\sigma} at temperature \code{temp2}.
}
\description{
\code{SigmaTransformed} transforms the measured sigma value at a specific
temperature to an estimate of sigma at another temperature. This useful in
comparitive seed testing protocol to compare seed longevities among species
tested at different temperatures
\insertCite{probert_ecological_2009}{viabilitymetrics}. \loadmathjax
}
\details{
The transformation is based on the effect of temperature on seed longevity
(\mjseqn{\sigma}) (identified by storage experiment with constant moisture
content and varying temperature) which is as follows.

\mjsdeqn{\log\sigma = \beta - C_{H}t - C_{H}t^{2}}

Where, \mjseqn{C_{H}} and \mjseqn{C_{Q}} are the species-specific temperature
coefficients, \mjseqn{t} is the temperature and \mjseqn{\beta} is the
constant associated with moisture relations of seed longevity.
}
\examples{

SigmaTransformed(250, 60, 45)

}
\references{
\insertRef{ellis_improved_1980}{viabilitymetrics}

\insertRef{probert_ecological_2009}{viabilitymetrics}
}
\seealso{
\code{\link[viabilitymetrics]{Sigma}}
}
