% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet2dry.R
\name{dry2wet}
\alias{dry2wet}
\alias{wet2dry}
\alias{MoistureNomograph}
\title{Conversion between Moisture content (wet/fresh weight basis) and Moisture
content (dry weight basis)}
\usage{
dry2wet(mc)

wet2dry(mc)

MoistureNomograph(min, max, basis = c("wet", "dry"), horiz = FALSE)
}
\arguments{
\item{mc}{Moisture content.}

\item{min}{Minimum value of moisture content to be plotted in nomograph.}

\item{max}{Minimum value of moisture content to be plotted in nomograph.}

\item{basis}{The basis on which moisture content is estimated}

\item{horiz}{If \code{TRUE}, nomograph is plotted horizontally.}
}
\value{
For \code{wet2dry} and \code{dry2wet}, the converted moisture content
(\%).

For \code{MoistureNomograph}, the nomograph as an object of class
\code{ggplot}.
}
\description{
These functions convert from moisture content determined on wet/fresh weight
basis to equivalent value on dry weight basis and back.
\code{MoistureNomograph} plots the nomograph for these conversions.
\loadmathjax
}
\details{
Conversions between moisture content (\%) determined on wet weight basis
\mjseqn{MC_{wb}} and that on dry weight basis \mjseqn{MC_{db}} are computed
based on the formulae
\insertCite{cromarty_design_1982;textual}{viabilitymetrics} as follows.

\mjsdeqn{MC_{db} = \frac{100 \times MC_{wb}}{100-MC_{wb}}}

\mjsdeqn{MC_{wb} = \frac{100 \times MC_{db}}{100+MC_{db}}}

\mjsdeqn{MC_{wb} = \frac{w_{i}-w_{f}}{w_{i}}}

\mjsdeqn{MC_{db} = \frac{w_{i}-w_{f}}{w_{f}}}

Where, \mjseqn{w_{i}} is the initial weight and \mjseqn{w_{f}} is the final
weight.

If the moisture content (\code{mc}) for conversion is beyond limits (0-50 \%
for \code{wet2dry} and 0-100 \% for \code{dry2wet}), a warning is issued.
}
\examples{
#----------------------------------------------------------------------------
# Moisture content (wet basis) to moisture content (dry basis)
#----------------------------------------------------------------------------
wet2dry(25)
# Warning if moisture content is beyond limits (0-50 \%)
wet2dry(60)
wet2dry(-10)

#----------------------------------------------------------------------------
# Moisture content (dry basis) to moisture content (wet basis)
#----------------------------------------------------------------------------
dry2wet(30)
# Warning if moisture content is beyond limits (0-100 \%)
dry2wet(-10)
dry2wet(110)

#----------------------------------------------------------------------------
# Nomograph
#----------------------------------------------------------------------------
# Horizontal
MoistureNomograph(min = 0, max = 50, basis = "wet", horiz = TRUE)
MoistureNomograph(min = 0, max = 100, basis = "dry", horiz = TRUE)

# Vertical
MoistureNomograph(min = 0, max = 50, basis = "wet", horiz = FALSE)
MoistureNomograph(min = 0, max = 100, basis = "dry", horiz = FALSE)

# Nomograph is a "ggplot" object
nom <- MoistureNomograph(min = 0, max = 50, basis = "wet", horiz = TRUE)
library(ggplot2)
nom + geom_hline(aes(yintercept=30), colour = "red")
p <- "Scale for converting moisture content values\nbetween dry and wet basis"
cap <- "based on Cromarty et al., 1982"
nom + labs(title  = p, caption = cap) +
  theme(plot.title = element_text(hjust = 0.5))
}
\references{
\insertRef{cromarty_design_1982}{viabilitymetrics}
}
