% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withWE.R
\name{withWE}
\alias{withWE}
\title{Evaluate expression and capture all warnings and errors if any along with results}
\usage{
withWE(expr)
}
\arguments{
\item{expr}{The expression to be evaluated.}
}
\description{
Not exported. Strictly internal
}
\section{Warning(s)}{

Returns the value along with the warning message(s).
}

\section{Error}{

Returns NA as the value along with the error message.
}

\examples{

foo <- function(){
  warning("oops")
}

withWE(foo)

foo <- function(){
  warning("oops")
  warning("again oops")
  1
}

withWE(foo)

foo <- function(){
  warning("oops")
  log("a")
}

withWE(foo)

}
\keyword{internal}
