% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSigma.batch.R
\name{FitSigma.batch}
\alias{FitSigma.batch}
\title{Seed viability curve fitting to estimate multiple values of
\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} according to a grouping
variable}
\usage{
FitSigma.batch(data, group, ...)
}
\arguments{
\item{data}{A data frame with the seed viability data recorded periodically.
It should possess columns with data on \itemize{ \item Viability percentage
(to be indicated by the argument \code{viability.percent}), \item Sample
size (to be indicated by the argument \code{samp.size}), \item Storage
period (to be indicated by the argument \code{storage.period}) and \item
Grouping variable (to be indicated by the argument \code{group}). }}

\item{group}{The name of the column in \code{data} with grouping variable as
a character string.}

\item{\dots}{Arguments to be passed on to \code{FitSigma}.}
}
\value{
A list of class \code{FitSigma.batch} with the following components:
  \item{data}{A data frame with the data used for computing the models.}
  \item{models}{A data frame with the group-wise values of model parameters,
  \ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}} and
  \ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}} and the fit statistics.}
}
\description{
Fit seed viability/survival curve to estimate multiple values of the seed lot
constant (\ifelse{html}{\out{<i>K<sub>i</sub></i>}}{\eqn{K_{i}}}) and the
period to lose unit probit viability
(\ifelse{html}{\out{<i>&sigma;</i>}}{\eqn{\sigma}}) according to a grouping
variable.
}
\examples{

data(seedsurvival)
df <- seedsurvival[seedsurvival$mc == 7 & seedsurvival$temp == 25,
                   c("crop", "period", "rep",
                     "viabilitypercent", "sampsize")]

#----------------------------------------------------------------------------
# Generalised linear model with probit link function (without cv)
#----------------------------------------------------------------------------
model1a <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize", storage.period = "period",
                          probit.method = "glm")
model1a

#----------------------------------------------------------------------------
# Generalised linear model with probit link function (with cv)
#----------------------------------------------------------------------------
model1b <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize", storage.period = "period",
                          probit.method = "glm",
                          use.cv = TRUE, control.viability = 98)
model1b

#----------------------------------------------------------------------------
# Linear model after probit transformation (without cv)
#----------------------------------------------------------------------------
model2a <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize", storage.period = "period",
                          probit.method = "tflm")
model2a

#----------------------------------------------------------------------------
# Linear model after probit transformation (with cv)
#----------------------------------------------------------------------------
model2b <- FitSigma.batch(data = df, group = "crop",
                          viability.percent = "viabilitypercent",
                          samp.size = "sampsize", storage.period = "period",
                          probit.method = "tflm",
                          use.cv = TRUE, control.viability = 98)
model2b
}
\seealso{
\code{\link[viabilitymetrics]{FitSigma}}
}
