% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viabilityconstants.R
\docType{data}
\name{viabilityconstants}
\alias{viabilityconstants}
\title{Viability constants.}
\format{A data frame with 7 columns: \describe{ \item{Index}{A unique index
  value, called by the \code{vcindex} argument in different functions.}
  \item{Species}{The species for which the constants were estimated.}
  \item{Ke}{The species-specific moisture constant
  \ifelse{html}{\out{<i>K<sub>E</sub></i>}}{\eqn{K_{E}}}} \item{Cw}{The
  species-specific moisture constant
  \ifelse{html}{\out{<i>C<sub>W</sub></i>}}{\eqn{C_{W}}}} \item{Ch}{The
  species-specific temperature constant
  \ifelse{html}{\out{<i>C<sub>H</sub></i>}}{\eqn{C_{H}}}} \item{Cq}{The
  species-specific temperature constant
  \ifelse{html}{\out{<i>C<sub>Q</sub></i>}}{\eqn{C_{Q}}}}
  \item{Reference}{Source of the data.} }}
\source{
Primarily from Hong et al., 1996, updated with more species. The
  source for each record is mentioned in the \strong{Reference} column.
}
\usage{
viabilityconstants
}
\description{
A dataset of the known species-specific seed viability constants estimated
from the 'improved seed viability equation' of Ellis and Roberts (1980).
}
\details{
Various functions such as \code{\link{Sigma}}, \code{\link{StorageMC}} use
the viability constants from this dataset for computation through the
argument \code{vcindex}.
}
\note{
Viability constants derived from storage experiment at a single
  temperature are highlighted by "*" in species column. In such cases,
  \ifelse{html}{\out{<i>C<sub>H</sub></i>}}{\eqn{C_{H}}} and
  \ifelse{html}{\out{<i>C<sub>Q</sub></i>}}{\eqn{C_{Q}}} are the universal
  temperature coefficients (0.0329 and 0.000478 respectively).
}
\examples{

# Soybean
viabilityconstants[grepl("glycine", x = viabilityconstants$Species,
                         ignore.case = T),]
# Rice
viabilityconstants[grepl("oryza", x = viabilityconstants$Species,
                         ignore.case = T),]
# Wheat
viabilityconstants[grepl("triticum", x = viabilityconstants$Species,
                         ignore.case = T),]
\dontrun{
# View dataset
View(viabilityconstants)
}

}
\references{
\insertRef{ellis_improved_1980}{viabilitymetrics}

  \insertRef{hong_seed_1996}{viabilitymetrics}
}
\keyword{datasets}
